requirements.txt
asgiref==3.3.1
autopep8==1.5.5
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
